requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
mysqlclient==2.0.3
python-dotenv==0.19.0
qrcode==7.3
Pillow==8.3.1
gunicorn==20.1.0